+++ src/main/java/com/maestronic/gtfs/service/TtsService.java
@@ -3,7 +3,7 @@ package com.maestronic.gtfs.service;
 import org.springframework.stereotype.Service;
 
 @Service
-public class TTSService {
+public class TtsService {
 
     public void textToSpeech(String text) {