+++ src/main/java/com/maestronic/gtfs/service/VehicleConnectionService.java
@@ -181,15 +181,16 @@ public class VehicleConnectionService implements GlobalVariable {
         return gtfs;
     }
 
-    public String getVehicleMonitorConnection(String agency_id, String vehicle_id) throws Exception {
+    public String getVehicleMonitorConnection(String agency_id, String vehicle_id, Long approx) throws Exception {
         List<VehicleMonitoring> resultList;
+        long timestamp = approx == null ? 0 : timeService.currentTimeToUnix() + approx;
         if (vehicle_id == null) {
-            resultList = vehicleMonitoringRepository.findVehicleMonitoringByAgency(agency_id, timeService.currentTimeToUnix());
+            resultList = vehicleMonitoringRepository.findVehicleMonitoringByAgency(agency_id, timestamp);
         } else {
             resultList = vehicleMonitoringRepository.findVehicleMonitoringByParam(
                     agency_id,
                     vehicle_id,
-                    timeService.currentTimeToUnix());
+                    timestamp);
         }
 
         if (resultList == null || resultList.size() <= 0) {